+++ src/main/java/org/concord/energy3d/simulation/EnergyAnnualAnalysis.java
@@ -214,13 +214,13 @@ public class EnergyAnnualAnalysis extends Analysis {
 		String s = null;
 		int cost = -1;
 		if (selectedPart != null) {
-			cost = Cost.getInstance().getPartCost(selectedPart);
+			cost = (int) Cost.getInstance().getPartCost(selectedPart);
 			if (graph.type == Graph.SENSOR) {
 				SceneManager.getInstance().setSelectedPart(null);
 			} else {
 				s = selectedPart.toString().substring(0, selectedPart.toString().indexOf(')') + 1);
 				if (selectedPart instanceof Foundation) {
-					cost = Cost.getInstance().getTotalCost();
+					cost = (int) Cost.getInstance().getTotalCost();
 					s = s.replaceAll("Foundation", "Building");
 					if (selectedPart.getChildren().isEmpty()) {
 						JOptionPane.showMessageDialog(MainFrame.getInstance(), "There is no building on this foundation.", "No Building", JOptionPane.INFORMATION_MESSAGE);
