+++ src/main/java/org/concord/energy3d/simulation/MirrorDailyAnalysis.java
@@ -158,7 +158,7 @@ public class MirrorDailyAnalysis extends Analysis {
 		String title = "Daily Yield of All Mirrors";
 		if (selectedPart != null) {
 			if (selectedPart instanceof Mirror) {
-				cost = Cost.getInstance().getPartCost(selectedPart);
+				cost = (int) Cost.getInstance().getPartCost(selectedPart);
 				s = selectedPart.toString().substring(0, selectedPart.toString().indexOf(')') + 1);
 				title = "Daily Yield";
 			} else if (selectedPart instanceof Foundation || selectedPart.getTopContainer() instanceof Foundation) {
